PUBDEV-5378: Split TargetEncoding test based on data leakage prevention strategy. Moved everything related to TE into targetencoding package. Moved frame related tests to the FrameTest.java
