--- h2o-automl/src/test/java/ai/h2o/automl/AutoMLAllTestsSuite.java
@@ -1,15 +1,0 @@
-package ai.h2o.automl;
-
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-        TargetEncodingTest.class
-})
-
-public class AutoMLAllTestsSuite {
-    // the class remains empty,
-    // used only as a holder for the above annotations
-}
